requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\image - Copy.py: 8
# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\image.py: 8
Flask == 1.1.2

# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\image - Copy.py: 9
# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\image.py: 9
Pillow == 8.4.0

# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\convert_weights.py: 19
# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\image - Copy.py: 5
# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\image.py: 5
# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\utils.py: 20
numpy == 1.20.3

# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\image - Copy.py: 4
# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\image.py: 4
# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\utils.py: 21
# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\video.py: 23
opencv_python == 4.5.4.60

# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\image - Copy.py: 2
# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\image.py: 2
# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\utils.py: 19
# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\video.py: 20
# H:\Deep-Learning-master\Deep-Learning-master\YOLOv3_TF2\yolov3.py: 21
tensorflow_cpu == 2.7.0
